# core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/song": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Create song",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Song"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/song/{id}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "summary": "Update song",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Song"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get single song",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Song"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/songs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get all songs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Song"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Create multiple songs at once",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Song"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/songs/updated/{time}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get updated songs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Song"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Maker": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Melody": {
            "type": "object",
            "properties": {
                "melody": {
                    "type": "string"
                }
            }
        },
        "model.Song": {
            "type": "object",
            "properties": {
                "arrangers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Maker"
                    }
                },
                "composers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Maker"
                    }
                },
                "creator": {
                    "type": "string"
                },
                "dateCreated": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "extraInfo": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastUpdated": {
                    "type": "string"
                },
                "lyricists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Maker"
                    }
                },
                "melody": {
                    "$ref": "#/definitions/model.Melody"
                },
                "modifier": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "verses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Verse"
                    }
                }
            }
        },
        "model.Tag": {
            "type": "object",
            "properties": {
                "tag": {
                    "type": "string"
                }
            }
        },
        "model.Verse": {
            "type": "object",
            "properties": {
                "lyrics": {
                    "type": "string"
                }
            }
        }
    }
}